Estimate adult mortality with one survey/census from sibling survival data.

Follows IUSSP(2012) template implementation for estimate adult mortality from siblings data, based on Timæus, Zaba and Ali (2001). The function also include the possibility to match the closest model life table (not using Brass logit given a pattern as default method), for some specific family (if no family is set, then return results for all CD and UN). An option for considering HIV populations is using the Spectrum model (Stover and others, 2012).

siblings_one(
  prop_not_dead,
  age,
  sex_siblings = "f",
  date = NULL,
  mlt_data_input = NULL,
  mlt_family = "CD_West",
  brass_logit = FALSE,
  brass_logit_e0 = 60,
  brass_logit_5q0 = NULL,
  HIV_prev = NULL,
  HIV_art = NULL,
  e0_accept = c(20, 100),
  verbose = TRUE
)

Arguments

prop_not_dead

numeric vector. Population proportion of respondents with siblings alive, by age groups.

age

integer vector. Lower bound of age groups from census. Last age is assumed as lower age from the open age group.

sex_siblings

character. "m" for male siblings, "f" for female siblings. Default is "f" given that most of the data is for sisters.

date

Either a Date class object or an unambiguous character string in the format "YYYY-MM-DD". Reference date for the source (typically census or survey).

mlt_data_input

data.frame. If is NULL then model life tables is used, available from Morcast package. But specific pattern can be included (l_x must be included see examples).

mlt_family

character. Options: "CD_East", "CD_North", "CD_South", "CD_West", "UN_Chilean", "UN_Far_Eastern", "UN_General", "UN_Latin_American", "UN_South_Asian". If NULL returns for all.

brass_logit

logical. Doing a level smoothing with 1-parameter logit Brass.

brass_logit_e0

numeric. Life expectancy at birth when brass_logit is TRUE.

brass_logit_5q0

numeric. Find best level when brass_logit_e0 is NULL. If is not NULL, then implied level replaces brass_logit_e0.

HIV_prev

numeric. Estimates of population-based estimate of HIV prevalence among men by age. If some value is assigned, then will be assumed AIDS variation of the method.

HIV_art

numeric. Estimate of the proportion of men with infected female partner.

e0_accept

integer vector. Range acceptable when calculating the median of implied level by age (avoid non-possible extrapolations). By deafult between 20 and 100.

Examples

if (FALSE) { # \dontrun{
# Examples from IUUSP (2012) Tools for Demographic Estimation
# Bangladesh 2003, all brothers. "AM_Siblings_Indirect - Bangladesh_2_5.xlsx"
standard_iusssp <- c(-0.8899, -0.8600,-0.8202,-0.7801, -0.7371, -0.6902, -0.6345, -0.5682, -0.4824, -0.3737)
standard_iusssp <- data.frame(age = seq(15, 60, 5), lx = 1/(1+exp(2*standard_iusssp)))
iussp_all_brothers <- siblings_one(prop_not_dead = c(0.946425,0.943374,0.941222,0.935335,0.928475,0.882786,0.832533),
                                   age = seq(15, 45, 5),
                                   sex_siblings = "m",
                                   date = 2003.4,
                                   mlt_data_input = standard_iusssp,
                                   brass_logit = TRUE)
iussp_results <- c(0.417, 0.344, 0.288, 0.242, 0.285, 0.306)
round(iussp_all_brothers$adult_mort_index$q15_45 - iussp_results,3)

# Bangladesh 2003, all sisters. "AM_Siblings_Indirect - Bangladesh_2_5.xlsx"
iussp_all_sisters <- siblings_one(prop_not_dead = c(0.9708,0.9447,0.9313,0.9180,0.9010,0.8748,0.8189),
                                  age = seq(15, 45, 5),
                                  sex_siblings = "f",
                                  date = 2003.4,
                                  mlt_data_input = standard_iusssp,
                                  brass_logit = TRUE)
iussp_results <- c(0.411, 0.384, 0.345, 0.315, 0.301, 0.327)
round(iussp_all_sisters$adult_mort_index$q15_45 - iussp_results,3)
} # }